backend/requirements.txt
openai
python-dotenv
sqlalchemy
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
fastapi
uvicorn
python-multipart
sentence-transformers
chromadb
python-docx
PyPDF2
langchain-community
langgraph
numpy
pydantic
passlib[bcrypt]
python-jose[cryptography]
structlog
opentelemetry-distro
opentelemetry-instrumentation-fastapi
smtplib
scikit-learn
transformers
torch
accelerate
